Move FileTransfer definitions to CPP file
